fix hash code

+++ cache.cpp
@@ -2,7 +2,7 @@
 #include "cache.h"
 #include "unordered_map"
 
-unordered_map<long, int> cache;
+unordered_map<long long, int> cache;
 int cacheSize = 0;
 int cacheCount = 0;
 
@@ -11,19 +11,19 @@ void initCache(int size)
 	cacheSize = size;
 }
 
-bool containsSearchKey(long key)
+bool containsSearchKey(long long key)
 {
 	if (cache.find(key) != cache.end())
 		return true;
 	return false;
 }
 
-int getSearchValue(long key)
+int getSearchValue(long long key)
 {
 	return cache[key];
 }
 
-void addSearchEntry(long key, int value)
+void addSearchEntry(long long key, int value)
 {
 	if (cacheCount > cacheSize)
 		return;
