+++ cache.h
@@ -2,10 +2,10 @@
 
 void initCache(int size);
 
-bool containsSearchKey(long key);
+bool containsSearchKey(long long key);
 
-int getSearchValue(long key);
+int getSearchValue(long long key);
 
-void addSearchEntry(long key, int value);
+void addSearchEntry(long long key, int value);
 
 void cacheReset();
